% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/docker.R
\name{get.deps.from.description}
\alias{get.deps.from.description}
\title{Extract R package dependencies from DESCRIPTION file}
\usage{
get.deps.from.description(dir)
}
\arguments{
\item{dir}{Directory where DESCRIPTION file is located}
}
\description{
\code{get.deps.from.description} extracts R package dependencies
from the package DESCRIPTION file and writes \emph{packages.R} script to
install R package dependencies in the Docker image. See \strong{Details} for exact
usage
}
\details{
The function is used to generate \emph{packages.R} script that lists
R package dependencies that need to be installed in Docker image for running the
Jupyter notebook under Docker. \cr\cr
The user should use this function, only when the automatically generated \emph{packages.R}
script is incomplete or malformatted.
}

