% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/processTex.R
\name{process.vignette}
\alias{process.vignette}
\title{Convert vignettes into Jupyter notebooks.}
\usage{
process.vignette(filename)
}
\arguments{
\item{filename}{File name of the R documentation (.Rmd or .Rnw) to be converted.
Absolute path has to be prepended if the file is not in the working directory.
The file extension must be included in the file name.}
}
\description{
\code{process.vignette} converts R markdown, Sweave and knitr vignettes
into Jupyter notebook format and automatically produces a Dockerfile for running
the notebook in Docker virtual environment.
}
\details{
The function expects as input the file name of the .Rmd or .Rnw documentation
that needs to be converted to Jupyter notebook format. Absolute path has to be prepended
 if the file is not in the working directory.
The file extension must be included in the file name.
\cr\cr R markdown (.Rmd) vignettes are first rendered to Markdown Github and
subsequently converted with the \strong{notedown}
command line utility to .ipynb format.
The produced Jupyter notebook is scanned for dependencies using \code{get.dependencies}
function and an R script \emph{packages.R} is produced. \emph{packages.R} and the
Jupyter notebook are used to make a Docker image for the vignette.
\cr\cr Note, however, that the Docker image has to be built first from the produced
Dockerfile. All generated files, i.e. Jupyter notebook, packages.R and Dockerfile
are written in the \strong{same} directory where the input file is located.
In order to build a Docker image for running the vignette you don't need
IPython Notebook installed on your machine. If you installed Docker,
simply run in command line \cr\cr
\code{$ docker build -t imagename .}
 \cr\cr in the same directory as the Dockerfile. After building the Docker image for
your Jupyter notebook, check the operating system-specific commands for running
Jupyter notebooks in Docker containers on the following
 web page \url{https://hub.docker.com/r/vladkim/rnaseq/}
 \cr\cr Similarly, Sweave and knitr vignettes are compiled to a TeX file, which
 is in turn converted to Markdown with \strong{pandoc}. The remaining processing
 steps are identical to those for R markdown vignettes.
}

