% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/docker.R
\name{produce.dockerfile}
\alias{produce.dockerfile}
\title{Produce a Docker image for Jupyter notebook.}
\usage{
produce.dockerfile(vignette)
}
\arguments{
\item{vignette}{Name of the Jupyter notebook file. Absolute path has to
be prepended if the notebook is not in R working directory.}
}
\description{
\code{produce.dockerfile} generates a Dockerfile
which can be used to build a Docker image for running Jupyter
notebook in Docker without the local installation of Jupyter notebook
}
\details{
The function scans the Jupyter notebook for R package dependencies and
produces a script \emph{packages.R} for installing these R packages in the Docker
image.
\cr\cr The Docker image is built from the generated
Dockerfile, which is written in the same directory as the Jupyter notebook.
In order to build a Docker image for running the vignette you don't need
IPython Notebook installed on your machine. If you installed Docker,
simply run in command line \cr\cr
\code{$ docker build -t imagename .}
 \cr\cr in the same directory as the Dockerfile. After building the Docker image for
your Jupyter notebook, check the operating system-specific commands for running
Jupyter notebooks in Docker containers on the following
 web page \url{https://hub.docker.com/r/vladkim/rnaseq/}
}

